+++ Sources/MapboxMaps/Style/StyleEncodable.swift
@@ -14,6 +14,12 @@ public extension StyleEncodable where Self: Encodable {
     func jsonObject() throws -> [String: Any] {
         try jsonObject(userInfo: [:])
     }
+
+    /// Encodes the receiver into a Dictionary representation.
+    /// - returns: A dictionary representation of the receiver.
+    func allStyleProperties() throws -> [String: Any] {
+        try allStyleProperties(userInfo: [:], shouldEncodeNilValues: false)
+    }
 }
 
 internal extension StyleEncodable where Self: Encodable {
@@ -30,5 +36,19 @@ internal extension StyleEncodable where Self: Encodable {
             throw TypeConversionError.invalidObject
         }
         return jsonObject
+    }
+
+    /// Encodes to a dictionary representation.
+    /// - Parameter userInfo: Provides some contexts for the encoding.
+    /// - returns: A dictionary representation of the object.
+    func allStyleProperties(
+        userInfo: [CodingUserInfoKey: Any],
+        shouldEncodeNilValues: Bool
+    ) throws -> [String: Any] {
+        let encoder = DictionaryEncoder()
+        encoder.userInfo = userInfo
+        encoder.shouldEncodeNilValues = shouldEncodeNilValues
+
+        return try encoder.encode(self)
     }
 }